refactor: extract activateUnitCommand helpers to reduce merge conflicts
Split the 663-line activateUnitCommand.ts into focused helper modules:
- elementalHelpers.ts: Element-to-key mapping and value arithmetic
- combatAccumulatorHelpers.ts: Combat accumulator add/remove operations
- nonCombatAbilityHelpers.ts: Heal, move, influence ability effects
- terrainModifierHelpers.ts: Terrain cost modifier application
- manaConsumptionHelpers.ts: Mana source consumption and restoration

Main file reduced from 663 to 246 lines. Each helper module can now be
edited independently, enabling parallel development without conflicts.
