+++ packages/core/src/engine/commands/units/helpers/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Unit activation helpers
+ *
+ * Re-exports all helper modules for unit ability activation.
+ */
+
+export * from "./elementalHelpers.js";
+export * from "./combatAccumulatorHelpers.js";
+export * from "./nonCombatAbilityHelpers.js";
+export * from "./terrainModifierHelpers.js";
+export * from "./manaConsumptionHelpers.js";
